add forecast prdective ai service

+++ src/main/java/com/greentechinnovators/entity/Data.java
@@ -12,7 +12,7 @@ public class Data {
     @Id
     private String id;
     private Double temp;
-    private Float humidity;
+    private Double humidity;
     private Double pressure; // (BMP280)
     private Double co2;      // (MQ-135)
     private Double gas;      // (MICS-5524)
@@ -20,7 +20,7 @@ public class Data {
     private Double light;    // (BH1750)
     private String device;   // (ESP32)
 
-    public Data(String id, Double temp, Float humidity , Double pressure, Double co2, Double gas, Double uv, Double light, String device) {
+    public Data(String id, Double temp, Double humidity, Double pressure, Double co2, Double gas, Double uv, Double light, String device) {
         this.id=id;
         this.temp = temp;
         this.humidity = humidity;
